Move loading autostart entries code into EntriesPage
